Got build button to turn on 'building mode' on and off and have dungeon area visually change.
