+++ src/actions/types.js
@@ -16,8 +16,9 @@ export const BAIT_HEROES = "BAIT_HEROES";
 
 export const UPDATE_PLAYER_TREASURE = "UPDATE_PLAYER_TREASURE";
 
-export const BUILD_DUNGEON = "BUILD_DUNGEON"
+export const BUILDING_MODE = "BUILDING_MODE"
 
+export const BUILD_DUNGEON = "BUILD_DUNGEON"
 
 // type for currently selected card
 export const SELECT_CARD = "SELECT_CARD"
