+++ src/components/css/Dungeon.css
@@ -45,6 +45,19 @@
     border-width: 5px;
     border-color: #534125;
 }
+.roomAreaBuilding{
+    height: 70%;
+    width: 80%;
+    background-color: #cccccc;
+    border-bottom-left-radius: 5px;
+    border-top-left-radius: 5px;
+    display: flex;
+    justify-content: flex-end;
+    align-items: center;
+    border-style: solid;
+    border-width: 5px;
+    border-color: #c78720;
+}
 .room{
     border-style: solid;
     border-width: 3px;
