+++ src/reducers/misc.js
@@ -1,22 +1,28 @@
 
 
 
-import { SELECT_CARD } from "../actions/types"
+import { SELECT_CARD, BUILDING_MODE } from "../actions/types"
 
 const initialState = {
     card: {},
-    className: ""
+    className: "",
+    buildingMode: false
 }
 
 const misc = (state = initialState, action) => {
 
     switch(action.type){
         case SELECT_CARD:
-            console.log(action)
             return {
                 ...state,
                 card: action.card,
                 className: action.className
+            }
+        case BUILDING_MODE:
+            console.log("buildingMode", !state.buildingMode)
+            return {
+                ...state,
+                buildingMode: !state.buildingMode // swaps state of buildingMode between true and false
             }
         default:
             return state
